moved to a provider pattern for DataTable
