+++ app/components/ui/data-table/index.ts
@@ -1,1 +1,2 @@
 export { default as DataTable } from "./DataTable.vue";
+export { default as DataTableProvider } from "./DataTableProvider.vue";
